fix(ci): satisfy script checks for npm integration
